integrated local/remote data source at repository
