+++ app/build.gradle
@@ -40,6 +40,7 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation("androidx.activity:activity-ktx:1.6.1")
     implementation("androidx.fragment:fragment-ktx:1.5.5")
+    implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.6.0"
     implementation 'com.google.android.material:material:1.8.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     // https://developers.google.com/android/guides/setup
