+++ app/src/main/java/com/freshdigitable/yttt/MainViewModel.kt
@@ -2,7 +2,10 @@ package com.freshdigitable.yttt
 
 import android.content.Intent
 import android.util.Log
-import androidx.lifecycle.*
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.asLiveData
+import androidx.lifecycle.viewModelScope
 import com.freshdigitable.yttt.data.AccountRepository
 import com.freshdigitable.yttt.data.AccountRepository.Companion.getNewChooseAccountIntent
 import com.freshdigitable.yttt.data.GoogleService
@@ -10,8 +13,7 @@ import com.freshdigitable.yttt.data.YouTubeLiveRepository
 import com.freshdigitable.yttt.data.model.LiveVideo
 import com.google.android.gms.common.GoogleApiAvailability
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.flow.*
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -21,10 +23,6 @@ class MainViewModel @Inject constructor(
     private val accountRepository: AccountRepository,
     private val googleService: GoogleService,
 ) : ViewModel() {
-    private val _onAir: MutableLiveData<List<LiveVideo>> = MutableLiveData()
-    val onAir: LiveData<List<LiveVideo>> = _onAir
-    private val _next: MutableLiveData<List<LiveVideo>> = MutableLiveData()
-    val next: LiveData<List<LiveVideo>> = _next
 
     fun getConnectionStatus(): Int = googleService.getConnectionStatusCode()
 
@@ -53,38 +51,24 @@ class MainViewModel @Inject constructor(
         viewModelScope.launch { fetchLiveStreams() }
     }
 
-    @OptIn(FlowPreview::class)
+    private val videos = liveRepository.videos
+    val onAir: LiveData<List<LiveVideo>> = videos.map { v ->
+        Log.d(TAG, "onair.runningFold: ${v.size}")
+        v.filter { it.isNowOnAir() }.sortedByDescending { it.actualStartDateTime }
+    }
+        .asLiveData(viewModelScope.coroutineContext)
+    val next: LiveData<List<LiveVideo>> = videos.map { v ->
+        Log.d(TAG, "upcoming.runningFold: ${v.size}")
+        v.filter { it.isUpcoming() }.sortedBy { it.scheduledStartDateTime }
+    }
+        .asLiveData(viewModelScope.coroutineContext)
+
     private suspend fun fetchLiveStreams() {
         val channelIds = liveRepository.fetchAllSubscribe().map { it.channelId }
         Log.d(TAG, "fetchSubscribeList: ${channelIds.size}")
-        val videos = channelIds.asFlow()
-            .map { liveRepository.fetchActivitiesList(it) }
-            .map { a -> a.map { it.videoId } }
-            .map { liveRepository.fetchVideoList(it) }
-            .flatMapConcat { v -> flow { v.forEach { emit(it) } } } /// ???
-            .filter { it.isLiveStream() }
-            .shareIn(viewModelScope, SharingStarted.Eagerly)
-
-        val onAirStream = videos.filter { it.isNowOnAir() }
-            .runningFold(listOf<LiveVideo>()) { acc, a ->
-                acc.toMutableList().apply { add(a) }
-                    .sortedBy { it.actualStartDateTime }
-            }
-        viewModelScope.launch {
-            onAirStream.collect {
-                _onAir.postValue(it)
-            }
-        }
-
-        val nextStream = videos.filter { it.isUpcoming() }
-            .runningFold(listOf<LiveVideo>()) { acc, v ->
-                acc.toMutableList().apply { add(v) }
-                    .sortedBy { it.scheduledStartDateTime }
-            }
-        viewModelScope.launch {
-            nextStream.collect {
-                _next.postValue(it)
-            }
+        channelIds.forEach { c ->
+            val logs = liveRepository.fetchLiveChannelLogs(c)
+            liveRepository.fetchVideoList(logs.map { it.videoId })
         }
     }
