+++ app/src/main/java/com/freshdigitable/yttt/data/model/LiveChannel.kt
@@ -6,3 +6,8 @@ interface LiveChannel {
 
     data class Id(val value: String)
 }
+
+data class LiveChannelEntity(
+    override val id: LiveChannel.Id,
+    override val title: String,
+) : LiveChannel
